Use roaringbitmap instead of vec usize

+++ src/storage/store.rs
@@ -5,7 +5,7 @@ use super::record;
 
 pub struct RecordStore {
     id_store: Vec<Rc<record::Record>>,
-    hash_store: HashMap<Rc<record::Record>, usize>,
+    hash_store: HashMap<Rc<record::Record>, u32>,
 }
 
 impl RecordStore {
@@ -17,13 +17,13 @@ impl RecordStore {
         }
     }
 
-    pub fn add(&mut self, record: &Rc<record::Record>) -> Option<usize> {
+    pub fn add(&mut self, record: &Rc<record::Record>) -> Option<u32> {
         let result = self.hash_store.get(record);
         match result {
             Some(_record) => None,
             _ => {
                 self.id_store.push(record.clone());
-                let id = self.id_store.len() -1;
+                let id = (self.id_store.len() -1) as u32;
                 self.hash_store.insert(record.clone(), id);
                 Some(id)
             }
@@ -31,14 +31,14 @@ impl RecordStore {
         
     }
 
-    pub fn get(&self, id: usize) -> Option<&Rc<record::Record>> {
-        match self.id_store.get(id) {
+    pub fn get(&self, id: u32) -> Option<&Rc<record::Record>> {
+        match self.id_store.get(id as usize) {
             Some(x) => Some(x),
             None => None,
         }
     }
 
-    pub fn multi_get(&self, ids: Vec<usize>) -> Vec<&Rc<record::Record>> {
+    pub fn multi_get(&self, ids: Vec<u32>) -> Vec<&Rc<record::Record>> {
         ids.into_iter().filter_map(|id| self.get(id)).collect()
     }
 }
